+++ DebugMode/release-notes.md
@@ -1,6 +1,8 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## 1.5
+* Updated for Stardew Valley 1.2.
 
 ## 1.4.1
 * Updated to SMAPI 1.9.