+++ FastAnimations/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## 1.2
+* Updated for Stardew Valley 1.2.
+
 ## 1.1
 * Fixed errors with drinking animation.