+++ RotateToolbar/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## 1.1
+* Updated for Stardew Valley 1.2.
+
 ## 1.0.1
 * Fixed controller input.